+++ ChangeLog.rst
@@ -6,6 +6,11 @@ Features
 
 - Detect annotation and other options from the existing output files
 
+Bug Fixes
+~~~~~~~~~
+
+- Fix "prequ check --silent" not being silent on outdated txt files
+
 1.1.0
 -----